requirements.txt
tensorflow-cpu==2.3.0
streamlit==0.62.0
numpy==1.16.5
pillow==6.2.0
h5py==2.10.0
opencv-python-headless==4.0.1.23
